Forgot to add something before last commit
The reason I didn't want to determine who won by whose turn it was, by the way, is because I want to re-use this code so I can modify it to have a CPU play against you. And I assume I'll need to have the CPU distinguish between where the Xs and Os are. If I completely got rid of that for the win check, I'd just have to re-implement it anyway. Short code for the sake of short code isn't the end goal.